WS_3_7: Fix sleep timer code
It looks like I really messed up the sleep timer code in the original
commit when I tried to add the 'overnight long sleep'.

Rewrite that code, to cater for both big sleeps that start either
before or after midnight.

Whilst there, make the SleepDuration/WakeupTime/SleepTime variables
consts, which specifically in this new code allows the compiler to
drop out the case that we are obviously not using. In my local tests
this dropped ~100bytes from the image size (and probably saved us
a single bit of math/test in the code itself).

Fixes: #217

Signed-off-by: Graham Whaley <graham.whaley@gmail.com>